refactor: consolidate parent-child task management and enhance testing
- Removed outdated `parent_child_tests.rs` and `parent_child.rs` files to streamline the codebase.
- Introduced new test files for `ask`, `notes`, `timeout`, and `unit` to improve organization and coverage of parent-child task interactions.
- Enhanced the `tools.rs` file by removing unnecessary imports and ensuring clarity in task management functions.
- Updated the setup functions for better test initialization and consistency across test modules.
- Added comprehensive tests to validate the behavior of parent-child task interactions, including ask handling, note management, and timeout scenarios.
